refactor: MCP adapter 결과 조립 책임 분리
